feature: introduce Service Providers and break out legacy provider
