+++ give.php
@@ -43,606 +43,382 @@ use Give\Container\Container;
 use Give\Form\Templates;
 use Give\Route\Form as FormRoute;
 use Give\Controller\Form as FormRouteController;
+use Give\ServiceProviders\LegacyServiceProvider;
+use Give\ServiceProviders\ServiceProvider;
 
 // Exit if accessed directly.
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-if ( ! class_exists( 'Give' ) ) :
+/**
+ * Main Give Class
+ *
+ * @since 2.8.0 build in a service container
+ * @since 1.0
+ *
+ * @property-read Give_API $api
+ * @property-read Give_Async_Process $async_process
+ * @property-read Give_Comment $comment
+ * @property-read Give_DB_Donors $donors
+ * @property-read Give_DB_Donor_Meta $donor_meta
+ * @property-read Give_Emails $emails
+ * @property-read Give_Email_Template_Tags $email_tags
+ * @property-read Give_DB_Form_Meta $form_meta
+ * @property-read Give_Admin_Settings $give_settings
+ * @property-read Give_HTML_Elements $html
+ * @property-read Give_Logging $logs
+ * @property-read Give_DB_Logs $log_db
+ * @property-read Give_DB_Log_Meta $logmeta_db
+ * @property-read Give_Notices $notices
+ * @property-read Give_DB_Payment_Meta $payment_meta
+ * @property-read Give_Roles $roles
+ * @property-read FormRoute $routeForm
+ * @property-read Templates $templates
+ * @property-read Give_Scripts $scripts
+ * @property-read Give_DB_Sequential_Ordering $sequential_donation_db
+ * @property-read Give_Sequential_Donation_Number $seq_donation_number
+ * @property-read Give_Session $session
+ * @property-read Give_DB_Sessions $session_db
+ * @property-read Give_Tooltips $tooltips
+ *
+ * @mixin Container
+ */
+final class Give {
+	/**
+	 * Give Template Loader Object
+	 *
+	 * @since  1.0
+	 * @access public
+	 *
+	 * @var    Give_Template_Loader object
+	 */
+	public $template_loader;
 
 	/**
-	 * Main Give Class
+	 * Give No Login Object
 	 *
-	 * @since 2.8.0 build in a service container
-	 * @since 1.0
+	 * @since  1.0
+	 * @access public
 	 *
-	 * @property-read Give_API $api
-	 * @property-read Give_Async_Process $async_process
-	 * @property-read Give_Comment $comment
-	 * @property-read Give_DB_Donors $donors
-	 * @property-read Give_DB_Donor_Meta $donor_meta
-	 * @property-read Give_Emails $emails
-	 * @property-read Give_Email_Template_Tags $email_tags
-	 * @property-read Give_DB_Form_Meta $form_meta
-	 * @property-read Give_Admin_Settings $give_settings
-	 * @property-read Give_HTML_Elements $html
-	 * @property-read Give_Logging $logs
-	 * @property-read Give_DB_Logs $log_db
-	 * @property-read Give_DB_Log_Meta $logmeta_db
-	 * @property-read Give_Notices $notices
-	 * @property-read Give_DB_Payment_Meta $payment_meta
-	 * @property-read Give_Roles $roles
-	 * @property-read FormRoute $routeForm
-	 * @property-read Templates $templates
-	 * @property-read Give_Scripts $scripts
-	 * @property-read Give_DB_Sequential_Ordering $sequential_donation_db
-	 * @property-read Give_Sequential_Donation_Number $seq_donation_number
-	 * @property-read Give_Session $session
-	 * @property-read Give_DB_Sessions $session_db
-	 * @property-read Give_Tooltips $tooltips
-	 *
-	 * @mixin Container
+	 * @var    Give_Email_Access object
 	 */
-	final class Give {
-		/**
-		 * Give Template Loader Object
-		 *
-		 * @since  1.0
-		 * @access public
-		 *
-		 * @var    Give_Template_Loader object
-		 */
-		public $template_loader;
+	public $email_access;
 
-		/**
-		 * Give No Login Object
-		 *
-		 * @since  1.0
-		 * @access public
-		 *
-		 * @var    Give_Email_Access object
-		 */
-		public $email_access;
+	/**
+	 * Give_Stripe Object.
+	 *
+	 * @since  2.5.0
+	 * @access public
+	 *
+	 * @var Give_Stripe
+	 */
+	public $stripe;
 
-		/**
-		 * Give_Stripe Object.
-		 *
-		 * @since  2.5.0
-		 * @access public
-		 *
-		 * @var Give_Stripe
-		 */
-		public $stripe;
+	/**
+	 * @since 2.8.0
+	 *
+	 * @var Container
+	 */
+	private $container;
 
-		/**
-		 * @var Container
-		 */
-		private $container;
+	/**
+	 * @since 2.8.0
+	 *
+	 * @var array Array of Service Providers to load
+	 */
+	private $serviceProviders = [
+		LegacyServiceProvider::class,
+	];
 
-		/**
-		 * Give constructor.
-		 *
-		 * Sets up the Container to be used for managing all other instances and data
-		 *
-		 * @since 2.8.0
-		 */
-		public function __construct() {
-			$this->container = new Container();
+	/**
+	 * Give constructor.
+	 *
+	 * Sets up the Container to be used for managing all other instances and data
+	 *
+	 * @since 2.8.0
+	 */
+	public function __construct() {
+		$this->container = new Container();
+	}
+
+	/**
+	 * Bootstraps the Give Plugin
+	 *
+	 * @since 2.8.0
+	 */
+	public function boot() {
+		// PHP version
+		if ( ! defined( 'GIVE_REQUIRED_PHP_VERSION' ) ) {
+			define( 'GIVE_REQUIRED_PHP_VERSION', '5.6.0' );
 		}
 
-		/**
-		 * Bootstraps the Give Plugin
-		 *
-		 * @since 2.8.0
-		 */
-		public function boot() {
-			// PHP version
-			if ( ! defined( 'GIVE_REQUIRED_PHP_VERSION' ) ) {
-				define( 'GIVE_REQUIRED_PHP_VERSION', '5.6.0' );
-			}
+		// Bailout: Need minimum php version to load plugin.
+		if ( function_exists( 'phpversion' ) && version_compare( GIVE_REQUIRED_PHP_VERSION, phpversion(), '>' ) ) {
+			add_action( 'admin_notices', [ $this, 'minimum_phpversion_notice' ] );
 
-			// Bailout: Need minimum php version to load plugin.
-			if ( function_exists( 'phpversion' ) && version_compare( GIVE_REQUIRED_PHP_VERSION, phpversion(), '>' ) ) {
-				add_action( 'admin_notices', [ $this, 'minimum_phpversion_notice' ] );
-
-				return;
-			}
-
-			// Add compatibility notice for recurring and stripe support with Give 2.5.0.
-			add_action( 'admin_notices', [ $this, 'display_old_recurring_compatibility_notice' ] );
-
-			$this->setup_constants();
-			$this->includes();
-			$this->bindClasses();
-			$this->init_hooks();
-
-			do_action( 'give_loaded' );
+			return;
 		}
 
-		/**
-		 * A helper for loading legacy classes that do not use autoloading, then binding their instance
-		 * to the container.
-		 *
-		 * @since 2.8.0
-		 *
-		 * @param string $alias
-		 * @param string|Closure $class
-		 * @param string $includesPath
-		 * @param bool $singleton
-		 */
-		private function bindLegacyInstance( $alias, $class, $includesPath, $singleton = false ) {
-			require_once GIVE_PLUGIN_DIR . "includes/$includesPath";
+		// Add compatibility notice for recurring and stripe support with Give 2.5.0.
+		add_action( 'admin_notices', [ $this, 'display_old_recurring_compatibility_notice' ] );
 
-			if ( $class instanceof Closure ) {
-				$this->container->instance( $alias, $class() );
-			} elseif ( $singleton ) {
-				$this->container->instance( $alias, $class::get_instance() );
-			} else {
-				$this->container->instance( $alias, new $class() );
-			}
-		}
+		add_action( 'plugins_loaded', [ $this, 'init' ], 0 );
 
-		/**
-		 * Binds the initial classes to the service provider.
-		 *
-		 * @since 2.8.0
-		 */
-		private function bindClasses() {
-			$this->container->singleton( 'templates', Templates::class );
-			$this->container->singleton( 'routeForm', FormRoute::class );
-		}
+		$this->setup_constants();
 
-		/**
-		 * Hook into actions and filters.
-		 *
-		 * @since  1.8.9
-		 */
-		private function init_hooks() {
-			register_activation_hook( GIVE_PLUGIN_FILE, 'give_install' );
-			add_action( 'plugins_loaded', [ $this, 'init' ], 0 );
-		}
+		do_action( 'give_loaded' );
+	}
 
+	/**
+	 * Init Give when WordPress Initializes.
+	 *
+	 * @since 1.8.9
+	 */
+	public function init() {
 		/**
-		 * Init Give when WordPress Initializes.
+		 * Fires before the Give core is initialized.
 		 *
 		 * @since 1.8.9
 		 */
-		public function init() {
-			/**
-			 * Fires before the Give core is initialized.
-			 *
-			 * @since 1.8.9
-			 */
-			do_action( 'before_give_init' );
+		do_action( 'before_give_init' );
 
-			// Set up localization.
-			$this->load_textdomain();
+		// Set up localization.
+		$this->load_textdomain();
 
-			$this->bindLegacyInstance( 'roles', 'Give_Roles', 'class-give-roles.php' );
-			$this->bindLegacyInstance( 'give_settings', 'Give_Admin_Settings', 'admin/class-admin-settings.php' );
-			$this->bindLegacyInstance( 'api', 'Give_API', 'api/class-give-api.php' );
-			$this->bindLegacyInstance( 'emails', 'Give_Emails', 'emails/class-give-emails.php' );
-			$this->bindLegacyInstance( 'email_tags', 'Give_Email_Template_Tags', 'emails/class-give-email-tags.php' );
-			$this->bindLegacyInstance( 'html', 'Give_HTML_Elements', 'admin/class-give-html-elements.php', true );
-			$this->bindLegacyInstance( 'donors', 'Give_DB_Donors', 'database/class-give-db-donors.php' );
-			$this->bindLegacyInstance( 'donor_meta', 'Give_DB_Donor_Meta', 'database/class-give-db-donor-meta.php' );
-			$this->bindLegacyInstance( 'tooltips', 'Give_Tooltips', 'class-give-tooltips.php' );
-			$this->bindLegacyInstance( 'notices', 'Give_Notices', 'class-notices.php' );
-			$this->bindLegacyInstance( 'payment_meta', 'Give_DB_Payment_Meta', 'database/class-give-db-payment-meta.php' );
-			$this->bindLegacyInstance( 'log_db', 'Give_DB_Logs', 'database/class-give-db-logs.php' );
-			$this->bindLegacyInstance( 'logmeta_db', 'Give_DB_Log_Meta', 'database/class-give-db-logs-meta.php' );
-			$this->bindLegacyInstance( 'logs', 'Give_Logging', 'class-give-logging.php' );
-			$this->bindLegacyInstance( 'form_meta', 'Give_DB_Form_Meta', 'database/class-give-db-form-meta.php' );
-			$this->bindLegacyInstance( 'sequential_donation_db', 'Give_DB_Sequential_Ordering', 'database/class-give-db-sequential-ordering.php' );
-			$this->bindLegacyInstance( 'async_process', 'Give_Async_Process', 'class-give-async-process.php' );
-			$this->bindLegacyInstance( 'scripts', 'Give_Scripts', 'class-give-scripts.php' );
-			$this->bindLegacyInstance( 'seq_donation_number', 'Give_Sequential_Donation_Number', 'payments/class-give-sequential-donation-number.php', true );
-			$this->bindLegacyInstance( 'comment', 'Give_Comment', 'class-give-comment.php', true );
-			$this->bindLegacyInstance( 'session_db', 'Give_DB_Sessions', 'database/class-give-db-sessions.php' );
-			$this->bindLegacyInstance( 'session', 'Give_Session', 'class-give-session.php', true );
+		$this->bindClasses();
 
-			// Load form template
-			$this->templates->load();
+		$this->loadServiceProviders();
 
-			// Load routes.
-			$this->routeForm->init( new FormRouteController() );
+		// Load form template
+		$this->templates->load();
 
-			/**
-			 * Fire the action after Give core loads.
-			 *
-			 * @since 1.8.7
-			 *
-			 * @param Give class instance.
-			 *
-			 */
-			do_action( 'give_init', $this );
-		}
+		// Load routes.
+		$this->routeForm->init( new FormRouteController() );
 
 		/**
-		 * Throw error on object clone
+		 * Fire the action after Give core loads.
 		 *
-		 * The whole idea of the singleton design pattern is that there is a single
-		 * object, therefore we don't want the object to be cloned.
+		 * @since 1.8.7
 		 *
-		 * @since  1.0
-		 * @access protected
+		 * @param Give class instance.
 		 *
-		 * @return void
 		 */
-		public function __clone() {
-			// Cloning instances of the class is forbidden.
-			give_doing_it_wrong( __FUNCTION__, __( 'Cheatin&#8217; huh?', 'give' ) );
+		do_action( 'give_init', $this );
+	}
+
+	/**
+	 * Setup plugin constants
+	 *
+	 * @since  1.0
+	 * @access private
+	 *
+	 * @return void
+	 */
+	private function setup_constants() {
+		// Plugin version.
+		if ( ! defined( 'GIVE_VERSION' ) ) {
+			define( 'GIVE_VERSION', '2.7.1' );
 		}
 
-		/**
-		 * Disable unserializing of the class
-		 *
-		 * @since  1.0
-		 * @access protected
-		 *
-		 * @return void
-		 */
-		public function __wakeup() {
-			// Unserializing instances of the class is forbidden.
-			give_doing_it_wrong( __FUNCTION__, __( 'Cheatin&#8217; huh?', 'give' ) );
+		// Plugin Root File.
+		if ( ! defined( 'GIVE_PLUGIN_FILE' ) ) {
+			define( 'GIVE_PLUGIN_FILE', __FILE__ );
 		}
 
-		/**
-		 * Setup plugin constants
-		 *
-		 * @since  1.0
-		 * @access private
-		 *
-		 * @return void
-		 */
-		private function setup_constants() {
-			// Plugin version.
-			if ( ! defined( 'GIVE_VERSION' ) ) {
-				define( 'GIVE_VERSION', '2.7.1' );
-			}
-
-			// Plugin Root File.
-			if ( ! defined( 'GIVE_PLUGIN_FILE' ) ) {
-				define( 'GIVE_PLUGIN_FILE', __FILE__ );
-			}
-
-			// Plugin Folder Path.
-			if ( ! defined( 'GIVE_PLUGIN_DIR' ) ) {
-				define( 'GIVE_PLUGIN_DIR', plugin_dir_path( GIVE_PLUGIN_FILE ) );
-			}
-
-			// Plugin Folder URL.
-			if ( ! defined( 'GIVE_PLUGIN_URL' ) ) {
-				define( 'GIVE_PLUGIN_URL', plugin_dir_url( GIVE_PLUGIN_FILE ) );
-			}
-
-			// Plugin Basename aka: "give/give.php".
-			if ( ! defined( 'GIVE_PLUGIN_BASENAME' ) ) {
-				define( 'GIVE_PLUGIN_BASENAME', plugin_basename( GIVE_PLUGIN_FILE ) );
-			}
-
-			// Make sure CAL_GREGORIAN is defined.
-			if ( ! defined( 'CAL_GREGORIAN' ) ) {
-				define( 'CAL_GREGORIAN', 1 );
-			}
+		// Plugin Folder Path.
+		if ( ! defined( 'GIVE_PLUGIN_DIR' ) ) {
+			define( 'GIVE_PLUGIN_DIR', plugin_dir_path( GIVE_PLUGIN_FILE ) );
 		}
 
-		/**
-		 * Include required files
-		 *
-		 * @since  1.0
-		 * @access private
-		 *
-		 * @return void
-		 */
-		private function includes() {
-			global $give_options;
-
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-cache-setting.php';
-
-			/**
-			 * Load libraries.
-			 */
-			if ( ! class_exists( 'WP_Async_Request' ) ) {
-				include_once GIVE_PLUGIN_DIR . 'includes/libraries/wp-async-request.php';
-			}
-
-			if ( ! class_exists( 'WP_Background_Process' ) ) {
-				include_once GIVE_PLUGIN_DIR . 'includes/libraries/wp-background-process.php';
-			}
-
-			require_once GIVE_PLUGIN_DIR . 'includes/setting-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/country-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/template-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/misc-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/forms/functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/ajax-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/currency-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/price-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/user-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/frontend-donor-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/gateways/functions.php';
-
-			/**
-			 * Load plugin files
-			 */
-			$give_options = give_get_settings();
-
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-cron.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-async-process.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-cache.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/post-types.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/filters.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/api/class-give-api-v2.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-tooltips.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-notices.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-translation.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-license-handler.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/admin/class-give-html-elements.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-scripts.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-donate-form.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-meta.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-comments.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-comments-meta.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-donors.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-donor-meta.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-form-meta.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-sequential-ordering.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-logs.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-logs-meta.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-sessions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/database/class-give-db-payment-meta.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-donor.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-stats.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-session.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-logging.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-comment.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/class-give-donor-wall-widget.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/forms/widget.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/forms/class-give-forms-query.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/forms/template.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/shortcodes.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/formatting.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/error-tracking.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/login-register.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/plugin-compatibility.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/deprecated/deprecated-classes.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/deprecated/deprecated-functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/deprecated/deprecated-actions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/deprecated/deprecated-filters.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/process-donation.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/backward-compatibility.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/actions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/class-payment-stats.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/class-payments-query.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/class-give-payment.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/payments/class-give-sequential-donation-number.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/gateways/actions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/gateways/paypal-standard.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/gateways/offline-donations.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/gateways/manual.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/emails/class-give-emails.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/emails/class-give-email-tags.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/admin/emails/class-email-notifications.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/emails/functions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/emails/template.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/emails/actions.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/class-give-donors-query.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/class-give-donor-wall.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/class-give-donor-stats.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/backward-compatibility.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/donors/actions.php';
-
-			require_once GIVE_PLUGIN_DIR . 'includes/admin/upgrades/class-give-updates.php';
-
-			require_once GIVE_PLUGIN_DIR . 'blocks/load.php';
-
-			// Include API
-			require_once GIVE_PLUGIN_DIR . 'src/API/API.php';
-
-			// Include Views
-			require_once GIVE_PLUGIN_DIR . 'src/Views/Views.php';
-
-			if ( defined( 'WP_CLI' ) && WP_CLI ) {
-				require_once GIVE_PLUGIN_DIR . 'includes/class-give-cli-commands.php';
-			}
-
-			// Load file for frontend
-			if ( $this->is_request( 'frontend' ) ) {
-				require_once GIVE_PLUGIN_DIR . 'includes/frontend/class-give-frontend.php';
-			}
-
-			if ( $this->is_request( 'admin' ) || $this->is_request( 'wpcli' ) ) {
-				require_once GIVE_PLUGIN_DIR . 'includes/admin/class-give-admin.php';
-			}// End if().
-
-			require_once GIVE_PLUGIN_DIR . 'includes/actions.php';
-			require_once GIVE_PLUGIN_DIR . 'includes/install.php';
-
-			// This conditional check will add backward compatibility to older Stripe versions (i.e. < 2.2.0) when used with Give 2.5.0.
-			if (
-				! defined( 'GIVE_STRIPE_VERSION' ) ||
-				(
-					defined( 'GIVE_STRIPE_VERSION' ) &&
-					version_compare( GIVE_STRIPE_VERSION, '2.2.0', '>=' )
-				)
-			) {
-				require_once GIVE_PLUGIN_DIR . 'includes/gateways/stripe/class-give-stripe.php';
-			}
+		// Plugin Folder URL.
+		if ( ! defined( 'GIVE_PLUGIN_URL' ) ) {
+			define( 'GIVE_PLUGIN_URL', plugin_dir_url( GIVE_PLUGIN_FILE ) );
 		}
 
-		/**
-		 * Loads the plugin language files.
-		 *
-		 * @since  1.0
-		 * @access public
-		 *
-		 * @return void
-		 */
-		public function load_textdomain() {
-			// Set filter for Give's languages directory
-			$give_lang_dir = dirname( plugin_basename( GIVE_PLUGIN_FILE ) ) . '/languages/';
-			$give_lang_dir = apply_filters( 'give_languages_directory', $give_lang_dir );
-
-			// Traditional WordPress plugin locale filter.
-			$locale = is_admin() && function_exists( 'get_user_locale' ) ? get_user_locale() : get_locale();
-			$locale = apply_filters( 'plugin_locale', $locale, 'give' );
-
-			unload_textdomain( 'give' );
-			load_textdomain( 'give', WP_LANG_DIR . '/give/give-' . $locale . '.mo' );
-			load_plugin_textdomain( 'give', false, $give_lang_dir );
+		// Plugin Basename aka: "give/give.php".
+		if ( ! defined( 'GIVE_PLUGIN_BASENAME' ) ) {
+			define( 'GIVE_PLUGIN_BASENAME', plugin_basename( GIVE_PLUGIN_FILE ) );
 		}
 
-
-		/**
-		 *  Show minimum PHP version notice.
-		 *
-		 * @since  1.8.12
-		 * @access public
-		 */
-		public function minimum_phpversion_notice() {
-			// Bailout.
-			if ( ! is_admin() ) {
-				return;
-			}
-
-			$notice_desc  = '<p><strong>' . __(
-				'Your site could be faster and more secure with a newer PHP version.',
-				'give'
-			) . '</strong></p>';
-			$notice_desc .= '<p>' . __(
-				'Hey, we\'ve noticed that you\'re running an outdated version of PHP. PHP is the programming language that WordPress and GiveWP are built on. The version that is currently used for your site is no longer supported. Newer versions of PHP are both faster and more secure. In fact, your version of PHP no longer receives security updates, which is why we\'re sending you this notice.',
-				'give'
-			) . '</p>';
-			$notice_desc .= '<p>' . __(
-				'Hosts have the ability to update your PHP version, but sometimes they don\'t dare to do that because they\'re afraid they\'ll break your site.',
-				'give'
-			) . '</p>';
-			$notice_desc .= '<p><strong>' . __( 'To which version should I update?', 'give' ) . '</strong></p>';
-			$notice_desc .= '<p>' . __(
-				'You should update your PHP version to either 5.6 or to 7.0 or 7.1. On a normal WordPress site, switching to PHP 5.6 should never cause issues. We would however actually recommend you switch to PHP7. There are some plugins that are not ready for PHP7 though, so do some testing first. PHP7 is much faster than PHP 5.6. It\'s also the only PHP version still in active development and therefore the better option for your site in the long run.',
-				'give'
-			) . '</p>';
-			$notice_desc .= '<p><strong>' . __( 'Can\'t update? Ask your host!', 'give' ) . '</strong></p>';
-			$notice_desc .= '<p>' . sprintf(
-				__(
-					'If you cannot upgrade your PHP version yourself, you can send an email to your host. If they don\'t want to upgrade your PHP version, we would suggest you switch hosts. Have a look at one of the recommended %1$sWordPress hosting partners%2$s.',
-					'give'
-				),
-				sprintf(
-					'<a href="%1$s" target="_blank">',
-					esc_url( 'https://wordpress.org/hosting/' )
-				),
-				'</a>'
-			) . '</p>';
-
-			echo sprintf(
-				'<div class="notice notice-error">%1$s</div>',
-				wp_kses_post( $notice_desc )
-			);
-		}
-
-		/**
-		 * Display compatibility notice for Give 2.5.0 and Recurring 1.8.13 when Stripe premium is not active.
-		 *
-		 * @since 2.5.0
-		 *
-		 * @return void
-		 */
-		public function display_old_recurring_compatibility_notice() {
-			// Show notice, if incompatibility found.
-			if (
-				defined( 'GIVE_RECURRING_VERSION' )
-				&& version_compare( GIVE_RECURRING_VERSION, '1.9.0', '<' )
-				&& defined( 'GIVE_STRIPE_VERSION' )
-				&& version_compare( GIVE_STRIPE_VERSION, '2.2.0', '<' )
-			) {
-				$message = sprintf(
-					__(
-						'<strong>Attention:</strong> GiveWP 2.5.0+ requires the latest version of the Recurring Donations add-on to process payments properly with Stripe. Please update to the latest version add-on to resolve compatibility issues. If your license is active, you should see the update available in WordPress. Otherwise, you can access the latest version by <a href="%1$s" target="_blank">logging into your account</a> and visiting <a href="%1$s" target="_blank">your downloads</a> page on the GiveWP website.',
-						'give'
-					),
-					esc_url( 'https://givewp.com/wp-login.php' ),
-					esc_url( 'https://givewp.com/my-account/#tab_downloads' )
-				);
-
-				Give()->notices->register_notice(
-					[
-						'id'               => 'give-compatibility-with-old-recurring',
-						'description'      => $message,
-						'dismissible_type' => 'user',
-						'dismiss_interval' => 'shortly',
-					]
-				);
-			}
-		}
-
-		/**
-		 * What type of request is this?
-		 *
-		 * @since 2.4.0
-		 *
-		 * @param string $type admin, ajax, cron or frontend.
-		 *
-		 * @return bool
-		 */
-		private function is_request( $type ) {
-			switch ( $type ) {
-				case 'admin':
-					return is_admin();
-				case 'ajax':
-					return defined( 'DOING_AJAX' );
-				case 'cron':
-					return defined( 'DOING_CRON' );
-				case 'frontend':
-					return ( ! is_admin() || defined( 'DOING_AJAX' ) ) && ! defined( 'DOING_CRON' ) && ! defined( 'REST_REQUEST' );
-				case 'wpcli':
-					return defined( 'WP_CLI' ) && WP_CLI;
-			}
-		}
-
-		/**
-		 * Magic properties are passed to the service container to retrieve the data.
-		 *
-		 * @since 2.8.0 retrieve from the service container
-		 * @since 2.7.0
-		 *
-		 * @param string $propertyName
-		 *
-		 * @return mixed
-		 * @throws Exception
-		 */
-		public function __get( $propertyName ) {
-			return $this->container->get( $propertyName );
-		}
-
-		/**
-		 * Magic methods are passed to the service container.
-		 *
-		 * @since 2.8.0
-		 *
-		 * @param $name
-		 * @param $arguments
-		 *
-		 * @return mixed
-		 */
-		public function __call( $name, $arguments ) {
-			return call_user_func_array( [ $this->container, $name ], $arguments );
+		// Make sure CAL_GREGORIAN is defined.
+		if ( ! defined( 'CAL_GREGORIAN' ) ) {
+			define( 'CAL_GREGORIAN', 1 );
 		}
 	}
 
-endif; // End if class_exists check
+	/**
+	 * Loads the plugin language files.
+	 *
+	 * @since  1.0
+	 * @access public
+	 *
+	 * @return void
+	 */
+	public function load_textdomain() {
+		// Set filter for Give's languages directory
+		$give_lang_dir = dirname( plugin_basename( GIVE_PLUGIN_FILE ) ) . '/languages/';
+		$give_lang_dir = apply_filters( 'give_languages_directory', $give_lang_dir );
 
+		// Traditional WordPress plugin locale filter.
+		$locale = is_admin() && function_exists( 'get_user_locale' ) ? get_user_locale() : get_locale();
+		$locale = apply_filters( 'plugin_locale', $locale, 'give' );
+
+		unload_textdomain( 'give' );
+		load_textdomain( 'give', WP_LANG_DIR . '/give/give-' . $locale . '.mo' );
+		load_plugin_textdomain( 'give', false, $give_lang_dir );
+	}
+
+
+	/**
+	 *  Show minimum PHP version notice.
+	 *
+	 * @since  1.8.12
+	 * @access public
+	 */
+	public function minimum_phpversion_notice() {
+		// Bailout.
+		if ( ! is_admin() ) {
+			return;
+		}
+
+		$notice_desc  = '<p><strong>' . __(
+			'Your site could be faster and more secure with a newer PHP version.',
+			'give'
+		) . '</strong></p>';
+		$notice_desc .= '<p>' . __(
+			'Hey, we\'ve noticed that you\'re running an outdated version of PHP. PHP is the programming language that WordPress and GiveWP are built on. The version that is currently used for your site is no longer supported. Newer versions of PHP are both faster and more secure. In fact, your version of PHP no longer receives security updates, which is why we\'re sending you this notice.',
+			'give'
+		) . '</p>';
+		$notice_desc .= '<p>' . __(
+			'Hosts have the ability to update your PHP version, but sometimes they don\'t dare to do that because they\'re afraid they\'ll break your site.',
+			'give'
+		) . '</p>';
+		$notice_desc .= '<p><strong>' . __( 'To which version should I update?', 'give' ) . '</strong></p>';
+		$notice_desc .= '<p>' . __(
+			'You should update your PHP version to either 5.6 or to 7.0 or 7.1. On a normal WordPress site, switching to PHP 5.6 should never cause issues. We would however actually recommend you switch to PHP7. There are some plugins that are not ready for PHP7 though, so do some testing first. PHP7 is much faster than PHP 5.6. It\'s also the only PHP version still in active development and therefore the better option for your site in the long run.',
+			'give'
+		) . '</p>';
+		$notice_desc .= '<p><strong>' . __( 'Can\'t update? Ask your host!', 'give' ) . '</strong></p>';
+		$notice_desc .= '<p>' . sprintf(
+			__(
+				'If you cannot upgrade your PHP version yourself, you can send an email to your host. If they don\'t want to upgrade your PHP version, we would suggest you switch hosts. Have a look at one of the recommended %1$sWordPress hosting partners%2$s.',
+				'give'
+			),
+			sprintf(
+				'<a href="%1$s" target="_blank">',
+				esc_url( 'https://wordpress.org/hosting/' )
+			),
+			'</a>'
+		) . '</p>';
+
+		echo sprintf(
+			'<div class="notice notice-error">%1$s</div>',
+			wp_kses_post( $notice_desc )
+		);
+	}
+
+	/**
+	 * Display compatibility notice for Give 2.5.0 and Recurring 1.8.13 when Stripe premium is not active.
+	 *
+	 * @since 2.5.0
+	 *
+	 * @return void
+	 */
+	public function display_old_recurring_compatibility_notice() {
+		// Show notice, if incompatibility found.
+		if (
+			defined( 'GIVE_RECURRING_VERSION' )
+			&& version_compare( GIVE_RECURRING_VERSION, '1.9.0', '<' )
+			&& defined( 'GIVE_STRIPE_VERSION' )
+			&& version_compare( GIVE_STRIPE_VERSION, '2.2.0', '<' )
+		) {
+			$message = sprintf(
+				__(
+					'<strong>Attention:</strong> GiveWP 2.5.0+ requires the latest version of the Recurring Donations add-on to process payments properly with Stripe. Please update to the latest version add-on to resolve compatibility issues. If your license is active, you should see the update available in WordPress. Otherwise, you can access the latest version by <a href="%1$s" target="_blank">logging into your account</a> and visiting <a href="%1$s" target="_blank">your downloads</a> page on the GiveWP website.',
+					'give'
+				),
+				esc_url( 'https://givewp.com/wp-login.php' ),
+				esc_url( 'https://givewp.com/my-account/#tab_downloads' )
+			);
+
+			Give()->notices->register_notice(
+				[
+					'id'               => 'give-compatibility-with-old-recurring',
+					'description'      => $message,
+					'dismissible_type' => 'user',
+					'dismiss_interval' => 'shortly',
+				]
+			);
+		}
+	}
+
+	/**
+	 * Load all the service providers to bootstrap the various parts of the application.
+	 *
+	 * @since 2.8.0
+	 */
+	private function loadServiceProviders() {
+		$providers = [];
+
+		foreach ( $this->serviceProviders as $serviceProvider ) {
+			if ( ! is_subclass_of( $serviceProvider, ServiceProvider::class ) ) {
+				throw new InvalidArgumentException( "$serviceProvider class must implement the ServiceProvider interface" );
+			}
+
+			/** @var ServiceProvider $serviceProvider */
+			$serviceProvider = new $serviceProvider();
+
+			$serviceProvider->register();
+
+			$providers[] = $serviceProvider;
+		}
+
+		foreach ( $providers as $serviceProvider ) {
+			$serviceProvider->boot();
+		}
+	}
+
+	/**
+	 * Register a Service Provider for bootstrapping
+	 *
+	 * @since 2.8.0
+	 *
+	 * @param $serviceProvider
+	 */
+	public function registerServiceProvider( $serviceProvider ) {
+		$this->serviceProviders[] = $serviceProvider;
+	}
+
+	/**
+	 * Magic properties are passed to the service container to retrieve the data.
+	 *
+	 * @since 2.8.0 retrieve from the service container
+	 * @since 2.7.0
+	 *
+	 * @param string $propertyName
+	 *
+	 * @return mixed
+	 * @throws Exception
+	 */
+	public function __get( $propertyName ) {
+		return $this->container->get( $propertyName );
+	}
+
+	/**
+	 * Magic methods are passed to the service container.
+	 *
+	 * @since 2.8.0
+	 *
+	 * @param $name
+	 * @param $arguments
+	 *
+	 * @return mixed
+	 */
+	public function __call( $name, $arguments ) {
+		return call_user_func_array( [ $this->container, $name ], $arguments );
+	}
+}
 
 /**
  * Start Give
@@ -654,10 +430,14 @@ endif; // End if class_exists check
  *
  * Example: <?php $give = Give(); ?>
  *
+ * @since 2.8.0 add parameter for quick retrieval from container
  * @since 1.0
+ *
+ * @param null $abstract Selector for data to retrieve from the service container
+ *
  * @return object|Give
  */
-function Give( $abstract = null ) { // phpcs:ignore
+function Give( $abstract = null ) {
 	static $instance = null;
 
 	if ( $instance === null ) {
